Fix a longstanding bug, some memory leaks
Fix a bug where on iOS6 collection table view would crash when
deallocated, due to obj_release being called on it after it has
been deallocated. We solve the issue by forcing the runtime to
retain a reference to the UITableView instance.

Additionally, fix some memory leaks under iOS7 by preventing custom
code for cell selection styling from being executed on that version
of OS, since the goal of the code is to imitate iOS7 to begin with.