Properly handle deletion errors
Handle cases when user deletes a comment he/she has no access to.
The proper behavior seems to be to temporarily show the comment as deleted
until a response comes back from the server saying that we have failed to
delete the content, at which point we restore the comment cell to its initial visibility state.